Cleanup Date Time Row
